Added variables to Tile: their coordinates on the display for display purposes and for checking validity of moves. Might change this from coordinates on display to coordinates on board. While one might make more sense, the other might be less computationally intensive. I'll probably end up changing it.
As for changes in the GUI: You can see what square you're about to place your piece in. I was going to add this further into the project, but I felt like it today so I did it. They're also supposed to glow green whenever the setup of your tiles are a possible move (not taking into account incorrect words). The sort method in here is incomplete. I also made them snap to grid, which was kind of tricky. There is a lot of math in the display, and I'm not sure if I should be doing that, but I'm VERY satisfied with the grid buffer approach as it is right now. After every move, I plan to paint each placed tile onto the grid. This will get very trick when it comes to undoing incorrect words. So like in real scrabble, you can play a word if it's incorrect, but your opponent is also allowed to challenge it. But after you've completed your turn and it comes for the AI to tell you you're wrong, it will be hard to remove the image of the tiles from the buffer. Maybe I'm making this too complex.

To add, there are a LOT of ownership issues, privacy issues, etc.. throughout the code. I will clean it up and make it all nice and theoretically okay later. Right now I'm just getting it to work.