Added more @TODOs to help code progression.
Board now correctly determines if tiles are in a line.

Dictionary is a blank class with a TODO in it.

Display now correctly snaps and places tiles (almost). I still have to get rid of artifacts after tile placement.

In main, I made it so that you cannot place a tile outside of the playing board. It also sorts the placed tiles by orientation after placement. This is important for determining whether tiles are in a valid placement for play.

Tile has a new toString method for debugging purposes. Also has a new equals override used for checking that tiles won't be placed on top of each other.

TileComparator is used so that when I have my arraylist of tiles, I can use the already present, list.sort(Comparator<T>) method to sort the elements.